fix: async middleware is equivalent to sync middleware - DIA-61984
