+++ tests/conftest.py
@@ -68,14 +68,6 @@ def environ(db_url, sqla_version_tuple, async_sqlalchemy_url):
         yield values
 
 
-@fixture(scope="session")
-def engine(environ):
-    from sqlalchemy import engine_from_config
-
-    engine = engine_from_config(environ, prefix="sqlalchemy_")
-    return engine
-
-
 @fixture(autouse=True)
 def tear_down(environ):
     from sqlalchemy.orm.session import close_all_sessions
@@ -88,6 +80,7 @@ def tear_down(environ):
     # reload fastapi_sqla to clear sqla deferred reflection mapping stored in Base
     importlib.reload(fastapi_sqla.models)
     importlib.reload(fastapi_sqla.sqla)
+    importlib.reload(fastapi_sqla.async_sqla)
     importlib.reload(fastapi_sqla)
